# Speed up op a bit