This stores the `SiteType`s made within `op` in a Vector instead of a Tuple, which is more appropriate since the number of SiteTypes/Tags is a dynamic value. This leads to:
```julia
julia> s = Index(2, "A,B,C,S=1/2")
(dim=2|id=499|"A,B,C,S=1/2")

julia> @btime op("Id", $s);
  1.634 μs (27 allocations: 2.17 KiB)
```
instead of the following on master:
```julia
julia> @btime op("Id", $s);
  2.352 μs (29 allocations: 2.17 KiB)
```

For a custom `op` overload it is even more pronounced:
```julia
julia> ITensors.op(::SiteType"S=1/2", ::OpName"Sz", s::Index) = itensor([1.0 0.0; 0.0 -1.0], s', s)

julia> @btime op("Sz", $s);
  893.650 ns (16 allocations: 1008 bytes)
```
compared to master:
```julia
julia> @btime op("Sz", $s);
  1.742 μs (20 allocations: 1.20 KiB)
```